feat: collect all auth reducers to auth reducer
