+++ src/services/reducers/index.js
@@ -3,17 +3,11 @@ import { combineReducers } from "redux";
 import allIngredientsReducer from "./allIngredients";
 import orderReducer from "./order";
 import ingredientModalReducer from "./ingredientModal";
-import registerReducer from "./register";
-import loginReducer from "./login";
-import logoutReducer from "./logout";
-import tokenReducer from "./token";
+import authreducer from "./auth";
 
 export const rootReducer = combineReducers({
   allIngredients: allIngredientsReducer,
   order: orderReducer,
   ingredientModal: ingredientModalReducer,
-  register: registerReducer,
-  login: loginReducer,
-  logout: logoutReducer,
-  token: tokenReducer,
+  auth: authreducer,
 });
